+++ src/Common/CHANGELOG.md
@@ -19,6 +19,7 @@ Unless otherwise noted, all changes made by [@jeremyknight-me](https://github.co
 
 - Added AddWorkDays function to DateTimeHelper.
 - Added DoesImplement<T> to TypeExtensions.
+- Added GetAttribute<T> to EnumHelper.
 
 ### Changed
 - Updated email validation regex.